fixed a bug where you could get unlimited joker when keep inventory was on fixed a bug where you had no item above your head after rejoining the server in team mode fixed a bug in team mode where the suffix was empty after rejoining

+++ src/main/java/de/veroxar/forceItemBattle/ForceItemBattle.java
@@ -45,8 +45,12 @@ public final class ForceItemBattle extends JavaPlugin {
 
     @Override
     public void onEnable() {
-        worldManager.deleteWorld(world);
         initializeData();
+        if (data.getConfigs().getSettingsConfig().toFileConfiguration().getBoolean("settings.WorldReset")) {
+            worldManager.deleteWorld(world);
+            data.getConfigs().getSettingsConfig().toFileConfiguration().set("settings.WorldReset", false);
+            data.getConfigs().getSettingsConfig().saveConfiguration();
+        }
         initializeDefaultTeams();
         loadCommands();
         loadListeners();
@@ -159,8 +163,6 @@ public final class ForceItemBattle extends JavaPlugin {
         saveConfigs();
         data.getLogic().removeAllTasks();
     }
-
-
 
     public static Data getData() {
         return data;
